방 접속 중복 방지 유저에게서 연속적으로 요청이 들어올 경우를 대비하여, 또 접속이 불안정해서 꺼졌을 경우를 대비하여 방에 join할 때 유저의 joinedRoomName을 보고 기존 방연결을 해제시키며 접속하게 함

+++ nodejs/Services/game/gameRoomLeaveService.js
@@ -10,16 +10,18 @@ const gameRoomLeaveService = (socket, msg) => {
   const room = SingleObject.RoomRepository.findByRoomName(roomName);
   const user = SingleObject.UserRepository.findById(id);
 
-  try {
-    if (room.deleteUser(user) === 0) {
-      SingleObject.RoomRepository.delete(room.getRoomName());
-    } else {
-      gameRoomUpdateService(socket, roomName, id);
+  if (user.joinedRoomName !== null) {
+    try {
+      if (room.deleteUser(user) === 0) {
+        SingleObject.RoomRepository.delete(room.getRoomName());
+      } else {
+        gameRoomUpdateService(socket, roomName, id);
+      }
+      user.updateJoinedRoomName(null);
+      retMsg.status = "ok";
+    } catch (err) {
+      console.log(err);
     }
-    user.updateJoinedRoomName(null);
-    retMsg.status = "ok";
-  } catch (err) {
-    console.log(err);
   }
 
   retMsg = JSON.stringify(retMsg);
